Configtool: set reasonable default values.
Especially, turn off display and such stuff by default. What to
set there can't be figured automatically, so we have to rely on
the user for turning them on.

The ability to set default values was achieved just recently,
that's why so many long existing values wanted to be changed.